367. 有效的完全平方数

+++ leetcode/src/main/java/com/leetcode/tag/must1/five/IsPerfectSquare.java
@@ -9,11 +9,11 @@ public class IsPerfectSquare {
             if (num < 2) {
                 return true;
             }
-            int left = 2;
-            int right = num / 2;
+            long left = 2;
+            long right = num / 2;
 
             while (left <= right) {
-                int mid = left + (right - left) / 2;
+                long mid = left + (right - left) / 2;
                 if (mid * mid == num) {
                     return true;
                 }
